Add templates for word-meeting feature
...also refactor example content hooks:
The meeting content creator grew big enough to be moved into its own
module.

...also link the template to the committee container in the default content.
