+++ opengever/examplecontent/profiles/municipality_content/opengever_content/04_committees.json
@@ -5,7 +5,12 @@
         "title_de": "Sitzungen",
         "title_fr": "Séances",
         "protocol_template": "resolvePath::/vorlagen/sablon-template-1",
+        "protocol_header_template": "resolvePath::/vorlagen/sablon-template-6",
+        "protocol_suffix_template": "resolvePath::/vorlagen/sablon-template-7",
         "excerpt_template": "resolvePath::/vorlagen/sablon-template-2",
+        "excerpt_header_template": "resolvePath::/vorlagen/sablon-template-8",
+        "excerpt_suffix_template": "resolvePath::/vorlagen/sablon-template-9",
+        "ad_hoc_template": "resolvePath::/vorlagen/ad-hoc-template",
         "agendaitem_list_template": "resolvePath::/vorlagen/sablon-template-3",
         "toc_template": "resolvePath::/vorlagen/sablon-template-4",
         "paragraph_template": "resolvePath::/vorlagen/sablon-template-5",
